Find taxonomic alignments for a list of names

This function uses APC & APNI to find taxonomic alignments for a list of names.

align_taxa(
  original_name,
  output = NULL,
  fuzzy_matching = TRUE,
  max_distance_abs = 3,
  max_distance_rel = 0.2,
  ver = default_version()
)

Arguments

original_name

A list of names to query for taxonomic alignments.

output

(optional) The name of the file to save the results to.

fuzzy_matching

An option to turn off fuzzy matching.

max_distance_abs

The absolute distance in substitution space.

max_distance_rel

The relative distance in substitution space.

ver

The version number for the flora resource.

Value

A tibble with columns: original_name, cleaned_name, aligned_name, source, known, and checked.

See also

Examples

align_taxa(c("Poa annua", "Abies alba"), output = "taxa_alignments.csv")
#> Checking alignments of 2 taxa
#>   -> 0 names already matched; 0 names checked but without a match; 2 taxa yet to be checked
#>   -> checking for exact matches for 2 species
#> loading object `taxonomic_resources` into global environment
#> Warning: downloaded length 0 != reported length 9
#> Warning: cannot open URL 'https://github.com/traitecoevo/ausflora/releases/download/0.0.1.9000/apc.parquet': HTTP status was '404 Not Found'
#> Error in download.file(paste0("https://github.com/traitecoevo/ausflora/releases/download/",     version, "/apc.parquet"), destfile = tmp_pqt): cannot open URL 'https://github.com/traitecoevo/ausflora/releases/download/0.0.1.9000/apc.parquet'