Access Australian Plant Census Dataset

This function provides access to the Australian Plant Census dataset containing information about various species. The dataset can be loaded from a github for a stable file or from a URL for the most cutting-edge, but not stable version.

dataset_access_function(version = NULL, path = NULL, type = "stable")

Arguments

version

Version number. The default is NULL, which will load the most recent version of the dataset on your computer or the most recent version known to the package if you have never downloaded the data before. With `plant_lookup_del`, specifying `version=NULL` will delete all data sets.

path

Path to store the data at. If not given, `datastorr` will use `rappdirs` to find the best place to put persistent application data on your system. You can delete the persistent data at any time by running `mydata_del(NULL)` (or `mydata_del(NULL, path)` if you use a different path).

type

Type of dataset to access. The default is "stable", which loads the dataset from a github archived file. If set to "current", the dataset will be loaded from a URL which is the cutting edge version, but this may change at any time without notice.

Examples

# Load the most recent version of the dataset from a local file
dataset_access_function()
#> [1] "0.0.1.9000"
#> $APC
#> # A tibble: 109,622 × 37
#>    taxonID       nameT…¹ accep…² accep…³ nomen…⁴ nomIl…⁵ nomIn…⁶ taxon…⁷ proPa…⁸
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <lgl>  
#>  1 https://id.b… scient… https:… Planta… NA      f       f       accept… FALSE  
#>  2 https://id.b… scient… https:… Charop… NA      f       f       accept… FALSE  
#>  3 https://id.b… scient… https:… Charop… orth. … f       f       orthog… FALSE  
#>  4 https://id.b… scient… https:… Equise… orth. … f       f       orthog… FALSE  
#>  5 https://id.b… scient… https:… Equise… NA      f       f       taxono… FALSE  
#>  6 https://id.b… scient… https:… Equise… NA      f       f       accept… FALSE  
#>  7 https://id.b… scient… https:… Cycadi… NA      f       f       accept… FALSE  
#>  8 https://id.b… scient… https:… Cycada… NA      f       f       accept… FALSE  
#>  9 https://id.b… scient… https:… Cycada… NA      f       f       taxono… FALSE  
#> 10 https://id.b… scient… https:… Cycada… NA      f       f       taxono… FALSE  
#> # … with 109,612 more rows, 28 more variables: scientificName <chr>,
#> #   scientificNameID <chr>, canonicalName <chr>,
#> #   scientificNameAuthorship <chr>, parentNameUsageID <chr>, taxonRank <chr>,
#> #   taxonRankSortOrder <dbl>, kingdom <chr>, class <chr>, subclass <chr>,
#> #   family <chr>, taxonConceptID <chr>, nameAccordingTo <chr>,
#> #   nameAccordingToID <chr>, taxonRemarks <chr>, taxonDistribution <chr>,
#> #   higherClassification <chr>, firstHybridParentName <chr>, …
#> 
#> $APNI
#> # A tibble: 132,331 × 47
#>    scientificN…¹ nameT…² scien…³ scien…⁴ canon…⁵ canon…⁶ nameE…⁷ nomen…⁸ scien…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 https://id.b… scient… Acanth… <scien… Acanth… <scien… ebarba… NA      R.M.Ba…
#>  2 https://id.b… autonym Acanth… <scien… Acanth… <scien… ebract… NA      NA     
#>  3 https://id.b… scient… Acanth… <scien… Acanth… <scien… ebract… NA      Vahl   
#>  4 https://id.b… autonym Acanth… <scien… Acanth… <scien… ilicif… NA      NA     
#>  5 https://id.b… scient… Acanth… <scien… Acanth… <scien… subint… NA      Nees   
#>  6 https://id.b… scient… Acanth… <scien… Acanth… <scien… typica  nom. i… Domin  
#>  7 https://id.b… scient… Acanth… <scien… Acanth… <scien… ilicif… NA      L.     
#>  8 https://id.b… scient… Acanth… <scien… Acanth… <scien… mollis  NA      L.     
#>  9 https://id.b… scient… Acanth… <scien… Acanth… <scien… volubi… NA      Wall.  
#> 10 https://id.b… scient… Acanth… <scien… Acanth… <scien… Acanth… NA      L.     
#> # … with 132,321 more rows, 38 more variables: autonym <lgl>, hybrid <lgl>,
#> #   cultivar <lgl>, formula <lgl>, scientific <lgl>, nomInval <lgl>,
#> #   nomIlleg <lgl>, namePublishedIn <chr>, namePublishedInID <chr>,
#> #   namePublishedInYear <dbl>, nameInstanceType <chr>, nameAccordingToID <chr>,
#> #   nameAccordingTo <chr>, originalNameUsage <chr>, originalNameUsageID <chr>,
#> #   originalNameUsageYear <chr>, typeCitation <chr>, kingdom <chr>,
#> #   family <chr>, genericName <chr>, specificEpithet <chr>, …
#> 

# Load a specific version of the dataset from a local file
dataset_access_function(version = 1.0, path = "/path/to/dataset")
#> Error in normalizePath(path, mustWork = TRUE): path[1]="/path/to/dataset": No such file or directory

# Load the current version of the dataset from a URL
dataset_access_function(type = "current")
#> $APC
#> # A tibble: 109,624 × 37
#>    taxonID       nameT…¹ accep…² accep…³ nomen…⁴ nomIl…⁵ nomIn…⁶ taxon…⁷ proPa…⁸
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <lgl>  
#>  1 https://id.b… scient… https:… Planta… NA      f       f       accept… FALSE  
#>  2 https://id.b… scient… https:… Charop… NA      f       f       accept… FALSE  
#>  3 https://id.b… scient… https:… Charop… orth. … f       f       orthog… FALSE  
#>  4 https://id.b… scient… https:… Equise… NA      f       f       taxono… FALSE  
#>  5 https://id.b… scient… https:… Equise… orth. … f       f       orthog… FALSE  
#>  6 https://id.b… scient… https:… Equise… NA      f       f       accept… FALSE  
#>  7 https://id.b… scient… https:… Cycadi… NA      f       f       accept… FALSE  
#>  8 https://id.b… scient… https:… Cycada… NA      f       f       accept… FALSE  
#>  9 https://id.b… scient… https:… Cycada… NA      f       f       taxono… FALSE  
#> 10 https://id.b… scient… https:… Cycada… NA      f       f       taxono… FALSE  
#> # … with 109,614 more rows, 28 more variables: scientificName <chr>,
#> #   scientificNameID <chr>, canonicalName <chr>,
#> #   scientificNameAuthorship <chr>, parentNameUsageID <chr>, taxonRank <chr>,
#> #   taxonRankSortOrder <dbl>, kingdom <chr>, class <chr>, subclass <chr>,
#> #   family <chr>, taxonConceptID <chr>, nameAccordingTo <chr>,
#> #   nameAccordingToID <chr>, taxonRemarks <chr>, taxonDistribution <chr>,
#> #   higherClassification <chr>, firstHybridParentName <chr>, …
#> 
#> $APNI
#> # A tibble: 132,360 × 47
#>    scientificN…¹ nameT…² scien…³ scien…⁴ canon…⁵ canon…⁶ nameE…⁷ nomen…⁸ scien…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 https://id.b… scient… Acanth… <scien… Acanth… <scien… ebarba… NA      R.M.Ba…
#>  2 https://id.b… autonym Acanth… <scien… Acanth… <scien… ebract… NA      NA     
#>  3 https://id.b… scient… Acanth… <scien… Acanth… <scien… ebract… NA      Vahl   
#>  4 https://id.b… autonym Acanth… <scien… Acanth… <scien… ilicif… NA      NA     
#>  5 https://id.b… scient… Acanth… <scien… Acanth… <scien… subint… NA      Nees   
#>  6 https://id.b… scient… Acanth… <scien… Acanth… <scien… typica  nom. i… Domin  
#>  7 https://id.b… scient… Acanth… <scien… Acanth… <scien… ilicif… NA      L.     
#>  8 https://id.b… scient… Acanth… <scien… Acanth… <scien… mollis  NA      L.     
#>  9 https://id.b… scient… Acanth… <scien… Acanth… <scien… volubi… NA      Wall.  
#> 10 https://id.b… scient… Acanth… <scien… Acanth… <scien… Acanth… NA      L.     
#> # … with 132,350 more rows, 38 more variables: autonym <lgl>, hybrid <lgl>,
#> #   cultivar <lgl>, formula <lgl>, scientific <lgl>, nomInval <lgl>,
#> #   nomIlleg <lgl>, namePublishedIn <chr>, namePublishedInID <chr>,
#> #   namePublishedInYear <dbl>, nameInstanceType <chr>, nameAccordingToID <chr>,
#> #   nameAccordingTo <chr>, originalNameUsage <chr>, originalNameUsageID <chr>,
#> #   originalNameUsageYear <chr>, typeCitation <chr>, kingdom <chr>,
#> #   family <chr>, genericName <chr>, specificEpithet <chr>, …
#> 

# Delete all versions of the dataset
dataset_access_function(version = NULL, path = NULL)
#> [1] "0.0.1.9000"
#> $APC
#> # A tibble: 109,622 × 37
#>    taxonID       nameT…¹ accep…² accep…³ nomen…⁴ nomIl…⁵ nomIn…⁶ taxon…⁷ proPa…⁸
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <lgl>  
#>  1 https://id.b… scient… https:… Planta… NA      f       f       accept… FALSE  
#>  2 https://id.b… scient… https:… Charop… NA      f       f       accept… FALSE  
#>  3 https://id.b… scient… https:… Charop… orth. … f       f       orthog… FALSE  
#>  4 https://id.b… scient… https:… Equise… orth. … f       f       orthog… FALSE  
#>  5 https://id.b… scient… https:… Equise… NA      f       f       taxono… FALSE  
#>  6 https://id.b… scient… https:… Equise… NA      f       f       accept… FALSE  
#>  7 https://id.b… scient… https:… Cycadi… NA      f       f       accept… FALSE  
#>  8 https://id.b… scient… https:… Cycada… NA      f       f       accept… FALSE  
#>  9 https://id.b… scient… https:… Cycada… NA      f       f       taxono… FALSE  
#> 10 https://id.b… scient… https:… Cycada… NA      f       f       taxono… FALSE  
#> # … with 109,612 more rows, 28 more variables: scientificName <chr>,
#> #   scientificNameID <chr>, canonicalName <chr>,
#> #   scientificNameAuthorship <chr>, parentNameUsageID <chr>, taxonRank <chr>,
#> #   taxonRankSortOrder <dbl>, kingdom <chr>, class <chr>, subclass <chr>,
#> #   family <chr>, taxonConceptID <chr>, nameAccordingTo <chr>,
#> #   nameAccordingToID <chr>, taxonRemarks <chr>, taxonDistribution <chr>,
#> #   higherClassification <chr>, firstHybridParentName <chr>, …
#> 
#> $APNI
#> # A tibble: 132,331 × 47
#>    scientificN…¹ nameT…² scien…³ scien…⁴ canon…⁵ canon…⁶ nameE…⁷ nomen…⁸ scien…⁹
#>    <chr>         <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 https://id.b… scient… Acanth… <scien… Acanth… <scien… ebarba… NA      R.M.Ba…
#>  2 https://id.b… autonym Acanth… <scien… Acanth… <scien… ebract… NA      NA     
#>  3 https://id.b… scient… Acanth… <scien… Acanth… <scien… ebract… NA      Vahl   
#>  4 https://id.b… autonym Acanth… <scien… Acanth… <scien… ilicif… NA      NA     
#>  5 https://id.b… scient… Acanth… <scien… Acanth… <scien… subint… NA      Nees   
#>  6 https://id.b… scient… Acanth… <scien… Acanth… <scien… typica  nom. i… Domin  
#>  7 https://id.b… scient… Acanth… <scien… Acanth… <scien… ilicif… NA      L.     
#>  8 https://id.b… scient… Acanth… <scien… Acanth… <scien… mollis  NA      L.     
#>  9 https://id.b… scient… Acanth… <scien… Acanth… <scien… volubi… NA      Wall.  
#> 10 https://id.b… scient… Acanth… <scien… Acanth… <scien… Acanth… NA      L.     
#> # … with 132,321 more rows, 38 more variables: autonym <lgl>, hybrid <lgl>,
#> #   cultivar <lgl>, formula <lgl>, scientific <lgl>, nomInval <lgl>,
#> #   nomIlleg <lgl>, namePublishedIn <chr>, namePublishedInID <chr>,
#> #   namePublishedInYear <dbl>, nameInstanceType <chr>, nameAccordingToID <chr>,
#> #   nameAccordingTo <chr>, originalNameUsage <chr>, originalNameUsageID <chr>,
#> #   originalNameUsageYear <chr>, typeCitation <chr>, kingdom <chr>,
#> #   family <chr>, genericName <chr>, specificEpithet <chr>, …
#>