Get Available Versions of the Australian Plant Census and Australian Plant Names Index Datasets

This function returns a list of available versions of the Australian Plant Census and Australian Plant Names Index datasets. The versions can be either local or from the Github repository.

dataset_versions(local = TRUE, path = NULL)

dataset_version_current(local = TRUE, path = NULL)

Arguments

local

Logical indicating if local or Github versions should be polled. The default is TRUE, which means that only local versions will be checked. If set to FALSE, the function will check for the most recent Github version if there are no local versions available. Note that local versions take precedence over Github versions if there are conflicts.

path

Path to store the data at.

Examples

# Get available local versions of the Australian Plant Census and Australian Plant Names Index datasets
dataset_versions()
#> [1] "0.0.1.9000"

# Get available Github versions of the Australian Plant Census and Australian Plant Names Index datasets
dataset_versions(local = FALSE)
#> [1] "0.0.0.9000" "0.0.1.9000"

# Get available versions of the Australian Plant Census and Australian Plant Names Index datasets stored at a specific path
dataset_versions(path = "/path/to/dataset")
#> Error in normalizePath(path, mustWork = TRUE): path[1]="/path/to/dataset": No such file or directory