Check if a vector of species are native anywhere in Australia

This function checks if the given species is native anywhere in Australia. It creates a lookup table from taxonomic resources, and checks if the species is listed as native in that table. Note that this will not detect within Australia invasives, e.g. if a species is from Western Australia and is invasive on the east coast.

native_anywhere_in_australia(species, resources = load_taxonomic_resources())

Arguments

species

A character string representing the binomial or common name of the species.

resources

An optional list of taxonomic resources to use for the lookup. If not provided, the function will load default taxonomic resources using the `load_taxonomic_resources()` function.

Value

A tibble with two columns: `species`, which is the same as the input `species`, and `native_anywhere_in_aus`, a logical vector indicating whether each species is native anywhere in Australia.

See also

Examples

resources<-load_taxonomic_resources()
#> Loading resources...
#> ...done
native_anywhere_in_australia("Eucalyptus globulus")
#> Loading resources...
#> ...done
#> # A tibble: 1 × 2
#>   species             native_anywhere_in_aus
#>   <chr>               <lgl>                 
#> 1 Eucalyptus globulus TRUE                  
native_anywhere_in_australia(c("Eucalyptus globulus","Pinus radiata"), resources=resources)
#> # A tibble: 2 × 2
#>   species             native_anywhere_in_aus
#>   <chr>               <lgl>                 
#> 1 Eucalyptus globulus TRUE                  
#> 2 Pinus radiata       FALSE