Calculate state-level diversity for native, introduced, and more complicated species origins

This function calculates state-level diversity for native, introduced, and more complicated species origins based on the geographic data available in the current Australian Plant Census.

state_diversity_counts(
  state = c("NSW", "NT", "Qld", "WA", "ChI", "NSW", "SA", "Vic", "Tas", "ACT", "NI",
    "LHI", "MI", "HI", "MDI", "CoI", "CSI", "AR", "CaI"),
  type_of_data = c("stable", "current"),
  ver = "0.0.1.9000"
)

Arguments

state

A character string indicating the Australian state or territory to calculate the diversity for. Default is "NSW". Possible values are "NSW", "NT", "Qld", "WA", "ChI", "SA", "Vic", "Tas", "ACT", "NI", "LHI", "MI", "HI", "MDI", "CoI", "CSI", and "AR".

type_of_data

A character string indicating the type of data to use for the calculation. Default is "stable". Possible values are "stable" and "current".

ver

A character string indicating the version of the data to use for the calculation. Default is "0.0.1.9000".

Value

A tibble of diversity counts for the specified state or territory, including native, introduced, and more complicated species origins. The tibble has three columns: "origin" indicating the origin of the species, "state" indicating the Australian state or territory, and "num_species" indicating the number of species for that origin and state.

Examples

state_diversity_counts(state = "NSW", type_of_data = "current", ver = "0.0.1.9000")
#> Rows: 109624 Columns: 37
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (31): taxonID, nameType, acceptedNameUsageID, acceptedNameUsage, nomenc...
#> dbl   (1): taxonRankSortOrder
#> lgl   (3): nomIlleg, nomInval, proParte
#> dttm  (2): created, modified
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 7 × 3
#>   origin                            state num_species
#>   <chr>                             <chr> <table[1d]>
#> 1 doubtfully naturalised            NSW     93       
#> 2 formerly naturalised              NSW      8       
#> 3 native                            NSW   5958       
#> 4 native and doubtfully naturalised NSW      2       
#> 5 native and naturalised            NSW     34       
#> 6 naturalised                       NSW   1580       
#> 7 presumed extinct                  NSW      8