Access Australian Plant Census Dataset

This function provides access to the Australian Plant Census dataset containing information about various species. The dataset can be loaded from a github for a stable file or from a URL for the most cutting-edge, but not stable version.

dataset_access_function(
  version = default_version(),
  path = NULL,
  type = "stable"
)

Arguments

version

Version number. The default is NULL, which will load the most recent version of the dataset on your computer or the most recent version known to the package if you have never downloaded the data before. With `plant_lookup_del`, specifying `version=NULL` will delete all data sets.

path

Path to store the data at. If not given, `datastorr` will use `rappdirs` to find the best place to put persistent application data on your system. You can delete the persistent data at any time by running `mydata_del(NULL)` (or `mydata_del(NULL, path)` if you use a different path).

type

Type of dataset to access. The default is "stable", which loads the dataset from a github archived file. If set to "current", the dataset will be loaded from a URL which is the cutting edge version, but this may change at any time without notice.

Examples

# Load the current version of the dataset direct from the APC website
dataset_access_function(type = "current")
#> $APC
#> # A tibble: 109,633 × 37
#>    taxonID    nameType acceptedNameUsageID acceptedNameUsage nomenclaturalStatus
#>    <chr>      <chr>    <chr>               <chr>             <chr>              
#>  1 https://i… scienti… https://id.biodive… Plantae Haeckel   NA                 
#>  2 https://i… scienti… https://id.biodive… Charophyta Sachs  NA                 
#>  3 https://i… scienti… https://id.biodive… Charophyta Sachs  orth. var.         
#>  4 https://i… scienti… https://id.biodive… Equisetopsida C.… orth. var.         
#>  5 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  6 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  7 https://i… scienti… https://id.biodive… Cycadidae Pax     NA                 
#>  8 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#>  9 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> 10 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> # ℹ 109,623 more rows
#> # ℹ 32 more variables: nomIlleg <chr>, nomInval <chr>, taxonomicStatus <chr>,
#> #   proParte <lgl>, scientificName <chr>, scientificNameID <chr>,
#> #   canonicalName <chr>, scientificNameAuthorship <chr>,
#> #   parentNameUsageID <chr>, taxonRank <chr>, taxonRankSortOrder <dbl>,
#> #   kingdom <chr>, class <chr>, subclass <chr>, family <chr>,
#> #   taxonConceptID <chr>, nameAccordingTo <chr>, nameAccordingToID <chr>, …
#> 
#> $APNI
#> # A tibble: 132,477 × 47
#>    scientificNameID     nameType scientificName scientificNameHTML canonicalName
#>    <chr>                <chr>    <chr>          <chr>              <chr>        
#>  1 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  2 https://id.biodiver… autonym  Acanthus ebra… <scientific><name… Acanthus ebr…
#>  3 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  4 https://id.biodiver… autonym  Acanthus ilic… <scientific><name… Acanthus ili…
#>  5 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  6 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  7 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  8 https://id.biodiver… scienti… Acanthus moll… <scientific><name… Acanthus mol…
#>  9 https://id.biodiver… scienti… Acanthus volu… <scientific><name… Acanthus vol…
#> 10 https://id.biodiver… scienti… Acanthus L.    <scientific><name… Acanthus     
#> # ℹ 132,467 more rows
#> # ℹ 42 more variables: canonicalNameHTML <chr>, nameElement <chr>,
#> #   nomenclaturalStatus <chr>, scientificNameAuthorship <chr>, autonym <lgl>,
#> #   hybrid <lgl>, cultivar <lgl>, formula <lgl>, scientific <lgl>,
#> #   nomInval <lgl>, nomIlleg <lgl>, namePublishedIn <chr>,
#> #   namePublishedInID <chr>, namePublishedInYear <dbl>, nameInstanceType <chr>,
#> #   nameAccordingToID <chr>, nameAccordingTo <chr>, originalNameUsage <chr>, …
#> 

# Load the a stable version of the dataset 
dataset_access_function(version="0.0.2.9000",type = "stable")
#> $APC
#> # A tibble: 109,624 × 37
#>    taxonID    nameType acceptedNameUsageID acceptedNameUsage nomenclaturalStatus
#>  * <chr>      <chr>    <chr>               <chr>             <chr>              
#>  1 https://i… scienti… https://id.biodive… Plantae Haeckel   NA                 
#>  2 https://i… scienti… https://id.biodive… Charophyta Sachs  NA                 
#>  3 https://i… scienti… https://id.biodive… Charophyta Sachs  orth. var.         
#>  4 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  5 https://i… scienti… https://id.biodive… Equisetopsida C.… orth. var.         
#>  6 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  7 https://i… scienti… https://id.biodive… Cycadidae Pax     NA                 
#>  8 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#>  9 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> 10 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> # ℹ 109,614 more rows
#> # ℹ 32 more variables: nomIlleg <chr>, nomInval <chr>, taxonomicStatus <chr>,
#> #   proParte <lgl>, scientificName <chr>, scientificNameID <chr>,
#> #   canonicalName <chr>, scientificNameAuthorship <chr>,
#> #   parentNameUsageID <chr>, taxonRank <chr>, taxonRankSortOrder <dbl>,
#> #   kingdom <chr>, class <chr>, subclass <chr>, family <chr>,
#> #   taxonConceptID <chr>, nameAccordingTo <chr>, nameAccordingToID <chr>, …
#> 
#> $APNI
#> # A tibble: 132,385 × 47
#>    scientificNameID     nameType scientificName scientificNameHTML canonicalName
#>  * <chr>                <chr>    <chr>          <chr>              <chr>        
#>  1 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  2 https://id.biodiver… autonym  Acanthus ebra… <scientific><name… Acanthus ebr…
#>  3 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  4 https://id.biodiver… autonym  Acanthus ilic… <scientific><name… Acanthus ili…
#>  5 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  6 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  7 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  8 https://id.biodiver… scienti… Acanthus moll… <scientific><name… Acanthus mol…
#>  9 https://id.biodiver… scienti… Acanthus volu… <scientific><name… Acanthus vol…
#> 10 https://id.biodiver… scienti… Acanthus L.    <scientific><name… Acanthus     
#> # ℹ 132,375 more rows
#> # ℹ 42 more variables: canonicalNameHTML <chr>, nameElement <chr>,
#> #   nomenclaturalStatus <chr>, scientificNameAuthorship <chr>, autonym <lgl>,
#> #   hybrid <lgl>, cultivar <lgl>, formula <lgl>, scientific <lgl>,
#> #   nomInval <lgl>, nomIlleg <lgl>, namePublishedIn <chr>,
#> #   namePublishedInID <chr>, namePublishedInYear <dbl>, nameInstanceType <chr>,
#> #   nameAccordingToID <chr>, nameAccordingTo <chr>, originalNameUsage <chr>, …
#>