Load taxonomic resources

Loads taxonomic resources into the global environment. This function accesses taxonomic data from a dataset using the provided version number or the default version. The loaded data contains two lists: APC and APNI, which contain taxonomic information about plant species in Australia. The function creates several data frames by filtering and selecting data from the loaded lists.

load_taxonomic_resources(
  ver = default_version(),
  reload = FALSE,
  filetype = "parquet"
)

Arguments

ver

The version number of the dataset to use. Defaults to the default version.

reload

A logical indicating whether to reload the dataset from the data source. Defaults to FALSE.

filetype

type of file to download. parquet or csv

Value

The taxonomic resources data loaded into the global environment.

Examples

load_taxonomic_resources()
#> loading object `taxonomic_resources` into global environment
#> $APC
#> # A tibble: 109,624 × 37
#>    taxonID    nameType acceptedNameUsageID acceptedNameUsage nomenclaturalStatus
#>  * <chr>      <chr>    <chr>               <chr>             <chr>              
#>  1 https://i… scienti… https://id.biodive… Plantae Haeckel   NA                 
#>  2 https://i… scienti… https://id.biodive… Charophyta Sachs  NA                 
#>  3 https://i… scienti… https://id.biodive… Charophyta Sachs  orth. var.         
#>  4 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  5 https://i… scienti… https://id.biodive… Equisetopsida C.… orth. var.         
#>  6 https://i… scienti… https://id.biodive… Equisetopsida C.… NA                 
#>  7 https://i… scienti… https://id.biodive… Cycadidae Pax     NA                 
#>  8 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#>  9 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> 10 https://i… scienti… https://id.biodive… Cycadales Pers. … NA                 
#> # ℹ 109,614 more rows
#> # ℹ 32 more variables: nomIlleg <chr>, nomInval <chr>, taxonomicStatus <chr>,
#> #   proParte <lgl>, scientificName <chr>, scientificNameID <chr>,
#> #   canonicalName <chr>, scientificNameAuthorship <chr>,
#> #   parentNameUsageID <chr>, taxonRank <chr>, taxonRankSortOrder <dbl>,
#> #   kingdom <chr>, class <chr>, subclass <chr>, family <chr>,
#> #   taxonConceptID <chr>, nameAccordingTo <chr>, nameAccordingToID <chr>, …
#> 
#> $APNI
#> # A tibble: 132,385 × 47
#>    scientificNameID     nameType scientificName scientificNameHTML canonicalName
#>  * <chr>                <chr>    <chr>          <chr>              <chr>        
#>  1 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  2 https://id.biodiver… autonym  Acanthus ebra… <scientific><name… Acanthus ebr…
#>  3 https://id.biodiver… scienti… Acanthus ebra… <scientific><name… Acanthus ebr…
#>  4 https://id.biodiver… autonym  Acanthus ilic… <scientific><name… Acanthus ili…
#>  5 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  6 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  7 https://id.biodiver… scienti… Acanthus ilic… <scientific><name… Acanthus ili…
#>  8 https://id.biodiver… scienti… Acanthus moll… <scientific><name… Acanthus mol…
#>  9 https://id.biodiver… scienti… Acanthus volu… <scientific><name… Acanthus vol…
#> 10 https://id.biodiver… scienti… Acanthus L.    <scientific><name… Acanthus     
#> # ℹ 132,375 more rows
#> # ℹ 42 more variables: canonicalNameHTML <chr>, nameElement <chr>,
#> #   nomenclaturalStatus <chr>, scientificNameAuthorship <chr>, autonym <lgl>,
#> #   hybrid <lgl>, cultivar <lgl>, formula <lgl>, scientific <lgl>,
#> #   nomInval <lgl>, nomIlleg <lgl>, namePublishedIn <chr>,
#> #   namePublishedInID <chr>, namePublishedInYear <dbl>, nameInstanceType <chr>,
#> #   nameAccordingToID <chr>, nameAccordingTo <chr>, originalNameUsage <chr>, …
#> 
#> $genera_accepted
#> # A tibble: 3,094 × 37
#>    taxonID    nameType acceptedNameUsageID acceptedNameUsage nomenclaturalStatus
#>    <chr>      <chr>    <chr>               <chr>             <chr>              
#>  1 https://i… scienti… https://id.biodive… Cycas L.          NA                 
#>  2 https://i… scienti… https://id.biodive… Bowenia Hook.f.   NA                 
#>  3 https://i… scienti… https://id.biodive… Lepidozamia Regel NA                 
#>  4 https://i… scienti… https://id.biodive… Macrozamia Miq.   NA                 
#>  5 https://i… scienti… https://id.biodive… Equisetum L.      NA                 
#>  6 https://i… scienti… https://id.biodive… Isoetes L.        NA                 
#>  7 https://i… scienti… https://id.biodive… Huperzia Bernh.   NA                 
#>  8 https://i… scienti… https://id.biodive… Lycopodiella Hol… NA                 
#>  9 https://i… scienti… https://id.biodive… Lycopodium L.     NA                 
#> 10 https://i… scienti… https://id.biodive… Phlegmariurus Ho… NA                 
#> # ℹ 3,084 more rows
#> # ℹ 32 more variables: nomIlleg <chr>, nomInval <chr>, taxonomicStatus <chr>,
#> #   proParte <lgl>, scientificName <chr>, scientificNameID <chr>,
#> #   canonicalName <chr>, scientificNameAuthorship <chr>,
#> #   parentNameUsageID <chr>, taxonRank <chr>, taxonRankSortOrder <dbl>,
#> #   kingdom <chr>, class <chr>, subclass <chr>, family <chr>,
#> #   taxonConceptID <chr>, nameAccordingTo <chr>, nameAccordingToID <chr>, …
#> 
#> $`APC list (accepted)`
#> # A tibble: 30,538 × 6
#>    canonicalName     scientificName     taxonomicStatus ID    stripped_canonical
#>    <chr>             <chr>              <chr>           <chr> <chr>             
#>  1 Cycas angulata    Cycas angulata R.… accepted        http… cycas angulata    
#>  2 Cycas arenicola   Cycas arenicola K… accepted        http… cycas arenicola   
#>  3 Cycas armstrongii Cycas armstrongii… accepted        http… cycas armstrongii 
#>  4 Cycas arnhemica   Cycas arnhemica K… accepted        http… cycas arnhemica   
#>  5 Cycas badensis    Cycas badensis K.… accepted        http… cycas badensis    
#>  6 Cycas basaltica   Cycas basaltica C… accepted        http… cycas basaltica   
#>  7 Cycas brunnea     Cycas brunnea K.D… accepted        http… cycas brunnea     
#>  8 Cycas cairnsiana  Cycas cairnsiana … accepted        http… cycas cairnsiana  
#>  9 Cycas calcicola   Cycas calcicola M… accepted        http… cycas calcicola   
#> 10 Cycas canalis     Cycas canalis K.D… accepted        http… cycas canalis     
#> # ℹ 30,528 more rows
#> # ℹ 1 more variable: stripped_scientific <chr>
#> 
#> $`APC list (known names)`
#> # A tibble: 69,892 × 6
#>    canonicalName         scientificName taxonomicStatus ID    stripped_canonical
#>    <chr>                 <chr>          <chr>           <chr> <chr>             
#>  1 Cycas ser. Furfurace… Cycas ser. Fu… taxonomic syno… http… cycas furfuraceos…
#>  2 Cycas ser. Yorkianos… Cycas ser. Yo… taxonomic syno… http… cycas yorkianosae 
#>  3 Cycas ser. Cairnsian… Cycas ser. Ca… taxonomic syno… http… cycas cairnsianos…
#>  4 Cycas ser. Endemicos… Cycas ser. En… taxonomic syno… http… cycas endemicosae 
#>  5 Cycas ser. Angulatos… Cycas ser. An… taxonomic syno… http… cycas angulatosae 
#>  6 Cycas ser. Armstrong… Cycas ser. Ar… taxonomic syno… http… cycas armstrongia…
#>  7 Cycas ser. Calcicolo… Cycas ser. Ca… taxonomic syno… http… cycas calcicolosae
#>  8 Cycas brunnea         Cycas brunnea… pro parte misa… http… cycas brunnea     
#>  9 Cycas media f. inerm… Cycas media f… nomenclatural … http… cycas media inerm…
#> 10 Cycas media var. ine… Cycas media v… nomenclatural … http… cycas media inerm…
#> # ℹ 69,882 more rows
#> # ℹ 1 more variable: stripped_scientific <chr>
#> 
#> $`APNI names`
#> # A tibble: 131,847 × 6
#>    canonicalName         scientificName ID    taxonomicStatus stripped_canonical
#>    <chr>                 <chr>          <chr> <chr>           <chr>             
#>  1 (Dendrobium pugionif… (Dendrobium p… http… unplaced        dendrobium pugion…
#>  2 (Dockrillia pugionif… (Dockrillia p… http… unplaced        dockrillia pugion…
#>  3 (Eucalyptus globulus… (Eucalyptus g… http… unplaced        eucalyptus globul…
#>  4 (Eucalyptus globulus… (Eucalyptus g… http… unplaced        eucalyptus globul…
#>  5 (Eucalyptus johnston… (Eucalyptus j… http… unplaced        eucalyptus johnst…
#>  6 (Eucalyptus johnston… (Eucalyptus j… http… unplaced        eucalyptus johnst…
#>  7 (Eucalyptus urnigera… (Eucalyptus u… http… unplaced        eucalyptus urnige…
#>  8 (Eucalyptus urnigera… (Eucalyptus u… http… unplaced        eucalyptus urnige…
#>  9 Abacopteris           Abacopteris F… http… unplaced        abacopteris       
#> 10 Abacopteris           Abacopteris F… http… unplaced        abacopteris       
#> # ℹ 131,837 more rows
#> # ℹ 1 more variable: stripped_scientific <chr>
#>